refactor: EventService 분할 및 Event 패키지 내부 정리 #76
- EventService를 EventQueryService와 EventCommandService로 분리했습니다.
- EventStockProducerService, StockCheckJob에서 불필요한 주석, 내용을 삭제했습니다.
- TicketingErrorCode에 PAGE_ILLEGAL_ARGUMENT 에러를 추가했습니다.
