+++ src/main/java/inu/codin/codinticketingapi/domain/ticketing/exception/TicketingErrorCode.java
@@ -19,7 +19,8 @@ public enum TicketingErrorCode implements GlobalErrorCode {
     PASSWORD_INVALID(HttpStatus.BAD_REQUEST, "관리자 비밀번호가 맞지 않습니다."),
     UNAUTHORIZED_EVENT_UPDATE(HttpStatus.UNAUTHORIZED, "인증되지 않은 이벤트 업데이트 입니다."),
     EVENT_ALREADY_STARTED(HttpStatus.BAD_REQUEST, "이미 이벤트가 시작했습니다."),
-    INSUFFICIENT_TOTAL_STOCK(HttpStatus.BAD_REQUEST, "변경하려는 재고는 남은 재고보다 적을 수 없습니다.");
+    INSUFFICIENT_TOTAL_STOCK(HttpStatus.BAD_REQUEST, "변경하려는 재고는 남은 재고보다 적을 수 없습니다."),
+    PAGE_ILLEGAL_ARGUMENT(HttpStatus.BAD_REQUEST, "페이지 번호는 1 이상이어야 합니다.");
 
     private final HttpStatus httpStatus;
     private final String message;
